Update require, move mock file
