+++ modules/claims_api/lib/claims_api/v2/mock_aws_service.rb
